Restrict blocked users, message sleep and special command for delete

+++ bot/handlers/promo.py
@@ -4,8 +4,9 @@ from aiogram import Router, F
 from aiogram.types import Message, CallbackQuery
 from aiogram.fsm.context import FSMContext
 
+from bot.controllers.blocked_user import is_user_blocked
 from bot.controllers.promo import create_promo, promo_exists
-from bot.controllers.user import user_exists
+from bot.controllers.user import get_user, user_exists
 from bot.markups.inline_markups import channels_kb, promo_kb, register_kb, promo_callback_data
 from bot.misc import bot
 from bot.states import PromoStates
@@ -17,8 +18,11 @@ promo_router = Router()
 @promo_router.callback_query(lambda query: query.data == promo_callback_data)
 async def enter_promo(callback_query: CallbackQuery, state: FSMContext):
     message = callback_query.message
+    user = await get_user(message.from_user.id)
+    if user and await is_user_blocked(user.phone_number):
+        return
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     await message.delete()
     if await user_exists(callback_query.from_user.id):
         await message.answer(ENTER_PROMO_PHOTO_TEXT)
@@ -31,7 +35,7 @@ async def enter_promo(callback_query: CallbackQuery, state: FSMContext):
 @promo_router.message(PromoStates.photo)
 async def register_promo_photo(message: Message, state: FSMContext):
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     photo = message.photo
     if not photo:
         await message.answer(ENTER_PHOTO_TEXT)
@@ -46,7 +50,7 @@ async def register_promo_photo(message: Message, state: FSMContext):
 @promo_router.message(PromoStates.promoCode)
 async def register_promo_code(message: Message, state: FSMContext):
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     promo_code = message.text
     await state.update_data(code=promo_code)
     promo_data = await state.get_data()
@@ -55,7 +59,11 @@ async def register_promo_code(message: Message, state: FSMContext):
     if not promo_code_exists:
         new_promo = await create_promo(user_id=message.from_user.id, **promo_data)
         await message.answer(PROMO_SAVED_TEXT)
+        await bot.send_chat_action(message.chat.id, 'typing')
+        await sleep(0.2)
         await message.answer(SPECIAL_CODE_TEXT.format(new_promo.special_code), reply_markup=promo_kb())
+        await bot.send_chat_action(message.chat.id, 'typing')
+        await sleep(0.2)
         await message.answer(CHANNELS_TEXT, reply_markup=channels_kb())
         return await state.clear()
     await message.answer(PROMO_HAS_BEEN_USED, reply_markup=promo_kb())
