+++ bot/handlers/registration.py
@@ -4,7 +4,8 @@ from aiogram.fsm.context import FSMContext
 from aiogram.types import ReplyKeyboardRemove
 from asyncio import sleep
 
-from bot.controllers.user import create_user, user_exists
+from bot.controllers.blocked_user import is_user_blocked
+from bot.controllers.user import create_user, get_user, user_exists
 from bot.misc import bot
 from bot.states import RegistrationStates
 from bot.texts import *
@@ -18,8 +19,11 @@ registration_router = Router()
 @registration_router.callback_query(lambda query: query.data == register_callback_data)
 async def register_user(callback_query: CallbackQuery, state: FSMContext):
     message = callback_query.message
+    user = await get_user(message.from_user.id)
+    if user and await is_user_blocked(user.phone_number):
+        return
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     await message.delete()
     if await user_exists(callback_query.from_user.id):
         await message.answer(SIGNED_UP_TEXT.format(callback_query.from_user.full_name), reply_markup=ReplyKeyboardRemove())
@@ -31,7 +35,7 @@ async def register_user(callback_query: CallbackQuery, state: FSMContext):
 @registration_router.message(RegistrationStates.name)
 async def register_name(message: Message, state: FSMContext):
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     user_name = message.text
     await state.update_data(name=user_name)
     await message.answer(ENTER_PHONE_NUMBER_TEXT, reply_markup=contact_kb())
@@ -41,7 +45,7 @@ async def register_name(message: Message, state: FSMContext):
 @registration_router.message(RegistrationStates.phone)
 async def register_phone_number(message: Message, state: FSMContext):
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     user_phone_number = message.contact.phone_number
     user_phone_number = '+' + user_phone_number if not user_phone_number.startswith('+') else user_phone_number
     await state.update_data(phone_number=user_phone_number)
@@ -52,7 +56,7 @@ async def register_phone_number(message: Message, state: FSMContext):
 @registration_router.message(RegistrationStates.address)
 async def register_address(message: Message, state: FSMContext):
     await bot.send_chat_action(message.chat.id, 'typing')
-    await sleep(0.5)
+    await sleep(0.2)
     user_address = message.text
     await state.update_data(address=user_address)
     user_data = await state.get_data()
@@ -62,5 +66,7 @@ async def register_address(message: Message, state: FSMContext):
         user_data.get('phone_number'),
         user_data.get('address'),
     ))
+    await bot.send_chat_action(message.chat.id, 'typing')
+    await sleep(0.2)
     await message.answer(FOR_ENTER_PROMO_TEXT, reply_markup=promo_kb())
     await state.clear()
